+++ packages/ui-components/src/components/test.ts
@@ -1,0 +1,8 @@
+import { html } from 'lit';
+import { BaseElement } from '../base';
+
+export class TestElement extends BaseElement {
+  render() {
+    return html` <div>Test</div> `;
+  }
+}
