+++ packages/ui-components/src/exports/test.ts
@@ -1,0 +1,9 @@
+import { TestElement } from '../components/test';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'test-element': TestElement;
+  }
+}
+
+customElements.define('test-element', TestElement);
